Remove the friendly name attribute from PFX private keys.
Windows, when creating a PFX, uses the friendly name attribute on a private key to record
the key name at the time of export. On import this same name is used, if available.
However, during a parallel load of the PFX there's a race condition where it has been determined
that the key name is not in use but the key file has not yet been written. This allows parallel
readers (same process or different) to collide on the key container name, and later cleanup can
cause unpredictable results.

The friendly name stripping was done with OpenSSL, using the following algorithm:
* Validate MAC (check existing password)
* Unpack the PKCS#7 authsafes from the PKCS#12 structure
* Foreach authsafe
    * If the authtype is other than data, copy it as-is to the new authsafes list.
    * Otherwise (aka if pkcs7-data)
        * Unpack the PKCS#12 safebag from the PKCS#7 data blob
        * If the safebag's attribute collection contains 1.2.840.113549.1.9.20 then remove it.
        * Pack the safebag (modified or no) into a PKCS#7 data blob, append it to the new authsafes list.
* Regenerate the MAC using the same password (since the encrypted materials are still using the same key)

This approach was chosen over doing an ephemeral import followed by an export in Windows because OpenSSL
exposes enough primitives to do the surgery without re-interpretation of the metadata, so it's a more
minimal change to the datafiles than using Windows would be. The before and after can be diffed under ASN.1
rules to determine that only two changes happened: 1) removal of any friendly name attributes,
2) MAC regeneration.