[release/6.0 ]Rename namespace of Mono.Cecil to avoid clash
We hit a case where the ILStrip task assembly is IL-merged with a few other task assemblies on the Xamarin side which causes a clash because of trying to merge different Mono.Cecil versions.
Rename the Mono.Cecil used here to avoid that clash.

Backport of https://github.com/dotnet/runtime-assets/pull/169